refactor(Input) : 사용자 입력을 항상 String으로 받은 후, 필요에 맞추어 변환해주도록 함. 그리고 Utils에 존재한 변환 메소드를 View 패키지로 옮김
